+++ app/src/components/chessMessageComponent/chessMessage.css
@@ -1,0 +1,9 @@
+.chessMessage {
+    height: 20%;
+    display: flex;
+    color: white;
+    justify-content: center;
+    align-items: center;
+    font-size: 36px;
+    font-weight: bold;
+}
