+++ app/src/components/moveListComponent/moveList.css
@@ -1,0 +1,7 @@
+.history_item{
+    font-size: 25px;
+    display: flex;
+    justify-content: center;
+    margin-top: 5px;
+    color: white;
+}
