+++ app/src/components/sideMenuComponent/sideMenu.css
@@ -1,0 +1,8 @@
+.sideMenu {
+    height: 621px;
+    width: 309px;
+    position: relative;
+    border: 1px solid white;
+    left: 19px;
+}
+
