+++ app/src/components/sideMenuComponent/sideMenu.test.js
